requirements_xfgv2.txt
# xFGv2 Structural Levels Requirements

# Core dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytz>=2023.3

# Web interface
flask>=2.3.0
flask-login>=0.6.0
werkzeug>=2.3.0

# File monitoring and scheduling
schedule>=1.2.0
watchdog>=3.0.0

# Data validation
pydantic>=2.0.0
marshmallow>=3.20.0

# CSV processing
chardet>=5.0.0

# Logging and monitoring
structlog>=23.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced features
numpy>=1.24.0
scipy>=1.10.0